Clean up `sat/t_bat` data
