+++ config/saturn/t_bat_user_syms.txt
@@ -4,10 +4,3 @@ _func_060C1618 = 0x060C1618;
 _UpdateAnim = 0x060C175C;
 _CreateAdditionalBats = 0x060CF2E8;
 _UpdatePrimitives = 0x060CF410;
-_prim = 0x060D2754;
-_isAlive = 0x060D2758;
-_positions = 0x060D2798;
-_facingLeft = 0x060D27D8;
-_offsets = 0x060D27F8;
-_fade = 0x060D2818;
-_idx = 0x060D2828;
