--- src/saturn/t_bat/batanim.h
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#ifndef SATURN_T_BAT_ANIM_H
-#define SATURN_T_BAT_ANIM_H
-
-#include "sattypes.h"
-
-extern AnimationFrame g_DefaultBatAnimationFrame[14];
-extern AnimationFrame g_BatAlternateAnimationFrame[13];
-extern AnimationFrame g_BatFarFromTargetAnimationFrame[14];
-extern AnimationFrame g_BatCloseToTargetAnimationFrame[40];
-extern AnimationFrame g_BatHighVelocityAnimationFrame[2];
-extern AnimationFrame* g_BatAnimationFrames[5];
-
-#endif
