+++ src/saturn/t_bat/batbss.c
@@ -8,7 +8,14 @@ s32 g_CurrentRoomX BSS = 0;
 s32 g_CurrentRoomY BSS = 0;
 s16 g_BatFollowDeltaX BSS = 0;
 s16 g_BatFollowDeltaY BSS = 0;
-u8 g_BatRuntimePadding[0xDC] BSS = {0};
+Primitive* prim BSS = NULL;
+bool isAlive[16] BSS = {0};
+Point16 positions[16] BSS = {0};
+s16 facingLeft[16] BSS = {0};
+s16 offsets[16] BSS = {0};
+u8 fade[16] BSS = {0};
+s32 idx BSS = 0;
+s32 g_BatRuntimePadding BSS = 0;
 FamiliarStats s_BatStats BSS = {0};
 u32 g_BatStatsPadding BSS = 0;
 Point16 s_BatPathingPoints[4][16] BSS = {0};
