+++ src/saturn/t_bat/batbss.h
@@ -9,6 +9,13 @@ extern s32 g_CurrentRoomX;
 extern s32 g_CurrentRoomY;
 extern s16 g_BatFollowDeltaX;
 extern s16 g_BatFollowDeltaY;
+extern Primitive* prim;
+extern bool isAlive[16];
+extern Point16 positions[16];
+extern s16 facingLeft[16];
+extern s16 offsets[16];
+extern u8 fade[16];
+extern s32 idx;
 extern FamiliarStats s_BatStats;
 extern Point16 s_BatPathingPoints[4][16];
