--- src/saturn/t_bat/batevent.c
@@ -1,184 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "batevent.h"
-
-#define EVENT_NONE (-1)
-#define EVENT_RELIC(id) (0x40000000 | (id))
-#define EVENT_CASTLE_FLAG(id) (0x80000000 | (id))
-#define BAT_EVENT(                                                             \
-    servant, roomX, roomY, cameraX, cameraY, condition, delay, entity, params) \
-    {NULL,    0,         servant, roomX,  roomY,  cameraX,                     \
-     cameraY, condition, delay,   entity, params, 1}
-#define NULL_EVENT {NULL, 0, -1, -1, -1, -1, -1, -1, -1, -1, 0, 0}
-
-enum BatEventCondition {
-    BAT_EVENT_FORM_OF_MIST = 7,
-    BAT_EVENT_NO2_SECRET_WALL_OPEN = 32,
-    BAT_EVENT_CAT_SPIKE_ROOM_LIT = 66,
-    BAT_EVENT_CHI_DEMON_SWITCH = 80,
-    BAT_EVENT_RCHI_DEMON_SWITCH = 88,
-    BAT_EVENT_LIB_BOOKSHELF_SECRET = 112,
-    BAT_EVENT_NZ0_SECRET_WALL_OPEN = 129,
-    BAT_EVENT_NZ0_SECRET_FLOOR_OPEN = 130,
-    BAT_EVENT_NZ1_LOWER_WALL_OPEN = 161,
-    BAT_EVENT_NZ1_UPPER_WALL_OPEN = 162,
-    BAT_EVENT_NO4_SECRET_FLOOR_OPEN = 196,
-    BAT_EVENT_NO4_SECRET_WALL_OPEN = 198,
-};
-
-enum BatEventServant {
-    BAT_EVENT_SERVANT_BAT = 1,
-    BAT_EVENT_SERVANT_GHOST = 2,
-    BAT_EVENT_SERVANT_FAERIE = 3,
-    BAT_EVENT_SERVANT_DEMON = 4,
-    BAT_EVENT_SERVANT_SWORD = 5,
-    BAT_EVENT_SERVANT_YOUSEI = 6,
-    BAT_EVENT_SERVANT_NOSE_DEMON = 7,
-};
-
-enum BatEventEntity {
-    BAT_EVENT_ENTITY_FAERIE = 220,
-    BAT_EVENT_ENTITY_DEMON = 223,
-};
-
-s32 s_IsServantDestroyed = 0;
-u32 s_LastTargetedEntityIndex = 0;
-
-ServantEvent g_Events[49] = {
-    NULL_EVENT,
-    BAT_EVENT(BAT_EVENT_SERVANT_BAT, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 1,
-              BAT_EVENT_ENTITY_DEMON, 15),
-    BAT_EVENT(BAT_EVENT_SERVANT_GHOST, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 1,
-              BAT_EVENT_ENTITY_DEMON, 15),
-    BAT_EVENT(BAT_EVENT_SERVANT_DEMON, 49, 20, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_DEMON, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_DEMON, 32, 40, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CHI_DEMON_SWITCH), 8,
-              BAT_EVENT_ENTITY_DEMON, 1),
-    BAT_EVENT(BAT_EVENT_SERVANT_DEMON, 16, 21, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_DEMON, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_DEMON, -31, 20, 0, 772,
-              EVENT_CASTLE_FLAG(BAT_EVENT_RCHI_DEMON_SWITCH), 8,
-              BAT_EVENT_ENTITY_DEMON, 1),
-    BAT_EVENT(BAT_EVENT_SERVANT_DEMON, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 1,
-              BAT_EVENT_ENTITY_DEMON, 15),
-    BAT_EVENT(BAT_EVENT_SERVANT_NOSE_DEMON, 49, 20, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_DEMON, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_NOSE_DEMON, 32, 40, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CHI_DEMON_SWITCH), 8,
-              BAT_EVENT_ENTITY_DEMON, 1),
-    BAT_EVENT(BAT_EVENT_SERVANT_NOSE_DEMON, 16, 21, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_DEMON, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_NOSE_DEMON, -31, 20, 0, 772,
-              EVENT_CASTLE_FLAG(BAT_EVENT_RCHI_DEMON_SWITCH), 8,
-              BAT_EVENT_ENTITY_DEMON, 1),
-    BAT_EVENT(BAT_EVENT_SERVANT_NOSE_DEMON, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 1,
-              BAT_EVENT_ENTITY_DEMON, 15),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 49, 20, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_FAERIE, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 16, 21, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_FAERIE, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 48,
-              BAT_EVENT_ENTITY_FAERIE, 1),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 21, 22, 187, 4,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 60, 26, 0, 255,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 44, 18, 791, 772,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 10, 15, 988, 4,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 39, 37, 0, 516,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NO4_SECRET_FLOOR_OPEN), 96,
-              BAT_EVENT_ENTITY_FAERIE, 3),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 36, 27, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NO4_SECRET_WALL_OPEN), 256,
-              BAT_EVENT_ENTITY_FAERIE, 4),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 12, 32, 0, 516,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ0_SECRET_WALL_OPEN), 160,
-              BAT_EVENT_ENTITY_FAERIE, 5),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 12, 32, 0, 511,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ0_SECRET_FLOOR_OPEN), 32,
-              BAT_EVENT_ENTITY_FAERIE, 6),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 27, 21, 640, 260,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NO2_SECRET_WALL_OPEN), 64,
-              BAT_EVENT_ENTITY_FAERIE, 7),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 49, 18, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_LIB_BOOKSHELF_SECRET), 512,
-              BAT_EVENT_ENTITY_FAERIE, 8),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 50, 11, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ1_UPPER_WALL_OPEN), 32,
-              BAT_EVENT_ENTITY_FAERIE, 9),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 43, 10, 0, 255,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ1_LOWER_WALL_OPEN), 128,
-              BAT_EVENT_ENTITY_FAERIE, 10),
-    BAT_EVENT(BAT_EVENT_SERVANT_FAERIE, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 1,
-              BAT_EVENT_ENTITY_FAERIE, 15),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 49, 20, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_FAERIE, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 16, 21, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_FAERIE, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 48,
-              BAT_EVENT_ENTITY_FAERIE, 1),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 21, 22, 187, 4,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 60, 26, 0, 255,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 44, 18, 791, 772,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 10, 15, 988, 4,
-              EVENT_RELIC(BAT_EVENT_FORM_OF_MIST), 96, BAT_EVENT_ENTITY_FAERIE,
-              2),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 39, 37, 0, 516,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NO4_SECRET_FLOOR_OPEN), 96,
-              BAT_EVENT_ENTITY_FAERIE, 3),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 36, 27, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NO4_SECRET_WALL_OPEN), 256,
-              BAT_EVENT_ENTITY_FAERIE, 4),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 12, 32, 0, 516,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ0_SECRET_WALL_OPEN), 160,
-              BAT_EVENT_ENTITY_FAERIE, 5),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 12, 32, 0, 511,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ0_SECRET_FLOOR_OPEN), 32,
-              BAT_EVENT_ENTITY_FAERIE, 6),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 27, 21, 640, 260,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NO2_SECRET_WALL_OPEN), 64,
-              BAT_EVENT_ENTITY_FAERIE, 7),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 49, 18, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_LIB_BOOKSHELF_SECRET), 512,
-              BAT_EVENT_ENTITY_FAERIE, 8),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 50, 11, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ1_UPPER_WALL_OPEN), 32,
-              BAT_EVENT_ENTITY_FAERIE, 9),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 43, 10, 0, 255,
-              EVENT_CASTLE_FLAG(BAT_EVENT_NZ1_LOWER_WALL_OPEN), 128,
-              BAT_EVENT_ENTITY_FAERIE, 10),
-    BAT_EVENT(BAT_EVENT_SERVANT_YOUSEI, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 1,
-              BAT_EVENT_ENTITY_FAERIE, 15),
-    BAT_EVENT(BAT_EVENT_SERVANT_SWORD, 49, 20, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_FAERIE, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_SWORD, 16, 21, 0, 4, EVENT_NONE, 0,
-              BAT_EVENT_ENTITY_FAERIE, 0),
-    BAT_EVENT(BAT_EVENT_SERVANT_SWORD, 41, 50, 0, 4,
-              EVENT_CASTLE_FLAG(BAT_EVENT_CAT_SPIKE_ROOM_LIT), 1,
-              BAT_EVENT_ENTITY_FAERIE, 15),
-    NULL_EVENT,
-};
-
-s32 g_PlaySfxStep = 99;
-s16 g_EntityRanges[4] = {5, 7, 32, 63};
-ServantEvent* g_EventQueue = g_Events;
