--- src/saturn/t_bat/batevent.h
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#ifndef SATURN_T_BAT_EVENT_H
-#define SATURN_T_BAT_EVENT_H
-
-#include "sattypes.h"
-
-extern s32 s_IsServantDestroyed;
-extern u32 s_LastTargetedEntityIndex;
-extern ServantEvent g_Events[49];
-extern s32 g_PlaySfxStep;
-extern s16 g_EntityRanges[4];
-extern ServantEvent* g_EventQueue;
-
-#endif
