--- src/saturn/t_bat/batgfx.c
@@ -1,2 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "gen/batgfx.h"
