--- src/saturn/t_bat/batstat.c
@@ -1,16 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "batstat.h"
-
-BatSpriteData g_BatSpriteData[4] = {
-    {-4, -4, 8, 8, 0x144, 0x78, 8, 0, 16, 8},
-    {-4, -4, 8, 8, 0x144, 0x78, 120, 8, 128, 16},
-    {-4, -4, 8, 8, 0x144, 0x78, 228, 135, 236, 143},
-    {-4, -4, 8, 8, 0x144, 0x78, 80, 0, 88, 8},
-};
-
-BatAbilityValues g_BatAbilityStats[10] = {
-    {90, 64, 0, 128, 1}, {90, 96, 0, 128, 1}, {60, 128, 1, 96, 1},
-    {60, 160, 1, 96, 0}, {60, 192, 2, 64, 0}, {60, 224, 2, 64, 0},
-    {45, 256, 2, 32, 0}, {45, 288, 3, 32, 0}, {30, 320, 3, 16, 0},
-    {30, 352, 3, 16, 0},
-};
