--- src/saturn/t_bat/sprbank.c
@@ -1,40 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "sprdata.h"
-
-SaturnBatSpritePart* g_ServantSpriteParts[35] = {
-    (SaturnBatSpritePart*)&g_BatEmptySpritePart,
-    &g_BatSpriteParts[0],
-    &g_BatSpriteParts[1],
-    &g_BatSpriteParts[2],
-    &g_BatSpriteParts[3],
-    &g_BatSpriteParts[4],
-    &g_BatSpriteParts[5],
-    &g_BatSpriteParts[6],
-    &g_BatSpriteParts[7],
-    &g_BatSpriteParts[8],
-    &g_BatSpriteParts[9],
-    &g_BatSpriteParts[10],
-    &g_BatSpriteParts[11],
-    &g_BatSpriteParts[12],
-    &g_BatSpriteParts[13],
-    &g_BatSpriteParts[14],
-    &g_BatSpriteParts[15],
-    &g_BatSpriteParts[16],
-    &g_BatSpriteParts[17],
-    &g_BatSpriteParts[18],
-    &g_BatSpriteParts[19],
-    &g_BatSpriteParts[20],
-    &g_BatSpriteParts[21],
-    &g_BatSpriteParts[22],
-    &g_BatSpriteParts[23],
-    &g_BatSpriteParts[24],
-    &g_BatSpriteParts[25],
-    &g_BatSpriteParts[26],
-    &g_BatSpriteParts[27],
-    &g_BatSpriteParts[28],
-    &g_BatSpriteParts[29],
-    &g_BatSpriteParts[30],
-    &g_BatSpriteParts[31],
-    &g_BatSpriteParts[32],
-    &g_BatSpriteParts[33],
-};
