--- src/saturn/t_bat/sprdata.c
@@ -1,27 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "sattypes.h"
-#include "sprdata.h"
-
-u32 g_BatSpriteResourceHeader[2] = {0x104C4100, 0xE44B4100};
-
-u32 g_BatEmptySpritePart = 0;
-
-SaturnBatSpritePart g_BatSpriteParts[34] = {
-    {0x100, 0, 0x000, 6, -4, 0},  {0x100, 0, 0x100, 6, -4, 0},
-    {0x100, 0, 0x000, 6, -4, 1},  {0x100, 0, 0x100, 6, -4, 1},
-    {0x100, 0, 0x000, 2, -4, 2},  {0x100, 0, 0x100, 2, -4, 2},
-    {0x100, 0, 0x000, 1, -4, 3},  {0x100, 0, 0x100, 1, -4, 3},
-    {0x100, 0, 0x000, 2, 1, 4},   {0x100, 0, 0x100, 2, 1, 4},
-    {0x100, 0, 0x000, 3, 1, 5},   {0x100, 0, 0x100, 3, 1, 5},
-    {0x100, 0, 0x000, 3, 0, 6},   {0x100, 0, 0x100, 3, 0, 6},
-    {0x100, 0, 0x000, 3, 0, 7},   {0x100, 0, 0x100, 3, 0, 7},
-    {0x100, 0, 0x000, 3, -1, 8},  {0x100, 0, 0x100, 3, -1, 8},
-    {0x100, 0, 0x000, 3, -1, 9},  {0x100, 0, 0x100, 3, -1, 9},
-    {0x100, 0, 0x200, 3, -7, 10}, {0x100, 0, 0x200, 3, -5, 11},
-    {0x100, 0, 0x200, 3, -1, 12}, {0x100, 0, 0x200, 3, 5, 13},
-    {0x100, 0, 0x200, 3, 7, 14},  {0x100, 0, 0x200, 3, 9, 15},
-    {0x100, 0, 0x200, 3, 8, 16},  {0x100, 0, 0x200, 3, 7, 17},
-    {0x100, 0, 0x200, 3, 5, 18},  {0x100, 0, 0x200, 3, 1, 19},
-    {0x100, 0, 0x200, 1, 0, 20},  {0x100, 0, 0x200, 0, 0, 21},
-    {0x100, 0, 0x300, 3, 0, 22},  {0x100, 0, 0x300, 1, 0, 23},
-};
